requirements.txt
asgiref                   ==    3.8.1
Django                      ==  5.1.2
djangorestframework           ==3.15.2
djangorestframework-simplejwt ==5.3.1
drf-yasg                   ==   1.21.8
inflection                ==    0.5.1
packaging                 ==    24.1
pillow                    ==    11.0.0
pip                        ==   24.2
PyJWT                     ==    2.9.0
python-decouple            ==   3.8
pytz                       ==   2024.2
PyYAML                      ==  6.0.2
rest-framework-simplejwt   ==   0.0.2
simplejwt                 ==    2.0.1
sqlparse                   ==   0.5.1
typing                ==        3.7.4.3
uritemplate              ==     4.1.1
